Add AmazingFeature

+++ api/routes/books.py
@@ -1,14 +1,14 @@
 from typing import OrderedDict
-
-from fastapi import APIRouter, status
+from fastapi import APIRouter, status, HTTPException
 from fastapi.responses import JSONResponse
 
 from api.db.schemas import Book, Genre, InMemoryDB
 
 router = APIRouter()
 
+# Initialize the database with some books
 db = InMemoryDB()
-db.books = {
+db.books = OrderedDict({
     1: Book(
         id=1,
         title="The Hobbit",
@@ -30,8 +30,20 @@ db.books = {
         publication_year=1955,
         genre=Genre.FANTASY,
     ),
-}
+})
 
+@router.get(
+    "/", response_model=OrderedDict[int, Book], status_code=status.HTTP_200_OK
+)
+async def get_books() -> OrderedDict[int, Book]:
+    return db.get_books()
+
+@router.get("/{book_id}", response_model=Book)
+async def get_book(book_id: int):
+    book = db.get_book(book_id)
+    if book is None:
+        raise HTTPException(status_code=404, detail="Book not found")
+    return book
 
 @router.post("/", status_code=status.HTTP_201_CREATED)
 async def create_book(book: Book):
@@ -40,23 +52,18 @@ async def create_book(book: Book):
         status_code=status.HTTP_201_CREATED, content=book.model_dump()
     )
 
-
-@router.get(
-    "/", response_model=OrderedDict[int, Book], status_code=status.HTTP_200_OK
-)
-async def get_books() -> OrderedDict[int, Book]:
-    return db.get_books()
-
-
 @router.put("/{book_id}", response_model=Book, status_code=status.HTTP_200_OK)
 async def update_book(book_id: int, book: Book) -> Book:
+    if not db.get_book(book_id):
+        raise HTTPException(status_code=404, detail="Book not found")
     return JSONResponse(
         status_code=status.HTTP_200_OK,
         content=db.update_book(book_id, book).model_dump(),
     )
 
-
 @router.delete("/{book_id}", status_code=status.HTTP_204_NO_CONTENT)
 async def delete_book(book_id: int) -> None:
+    if not db.get_book(book_id):
+        raise HTTPException(status_code=404, detail="Book not found")
     db.delete_book(book_id)
     return JSONResponse(status_code=status.HTTP_204_NO_CONTENT, content=None)
